Finish playRound function
Finished the playRound function.  It will randomly generate an AI answer, and compare it to the user's answer that they entered into a prompt.  The function is very busy, and I would like to maybe make some helper functions that can do the comparing operations so that the playRound function doesn't have to do all of the heavy lifting.