Change "visible" property to "display" for menu elements
Changed the start menu and the points prompt to "display" in CSS so that
I could get the elements to dissapear entirely instead of being invisible
but taking up space.  This should make them easier to format with CSS,
but if they get finicky, remember that I can change them back to visible.