Add autoplay function and round history
Added an autoplay function that automatically selects a player's choice
and then plays each round on a loop for the given number of rounds.

Also added a history feature that gives the stats of each round and lists them.
I will be adding a button to hide these next.  Perhaps also turning them
into a table format would be ideal.