Add History Table, standardize naming conventions, fix bug
Added a history table that goes over the detail of each round.
Also fixed a bug that I found out BECAUSE of the table that was making
two different variables for the computer choice and confusing the
createRow function.
Also standardized some of the names of functions and parameters to
make it a little prettier.